feat(mme): 5G AMF additions #7565
Test case :
    1) Verified basic flows using UERANSIM
        a) Initial Registration
        b) PDUSession
        c) Deregistration
    2) Run all mme unit test cases

+++ lte/gateway/c/core/oai/tasks/nas5g/src/ies/M5GDNN.cpp
@@ -54,7 +54,7 @@ int DNNMsg::EncodeDNNMsg(
   IES_ENCODE_U8(buffer, encoded, dnn_message->len);
   MLOG(MDEBUG) << "Length = " << std::hex << int(dnn_message->len);
   std::copy(dnn_message->dnn.begin(), dnn_message->dnn.end(), buffer + encoded);
-  BUFFER_PRINT_LOG(buffer + encoded, (int) dnn_message->dnn.length());
+  BUFFER_PRINT_LOG(buffer + encoded, dnn_message->dnn.length());
   encoded = encoded + dnn_message->dnn.length();
 
   return encoded;
